FastDDS interop: IF there are stale ACKNACKs, do not emit scary errors and do not reset ack_base.

+++ src/network/udp_listener.rs
@@ -424,7 +424,7 @@ impl UDPListener {
       let mut spec_dst: Option<IpAddr> = None;
       for cmsg in msg.cmsgs()? {
         if let ControlMessageOwned::Ipv4PacketInfo(info) = cmsg {
-          ifindex = info.ipi_ifindex as u32;
+          ifindex = info.ipi_ifindex;
           let addr = Ipv4Addr::from(u32::from_be(info.ipi_spec_dst.s_addr));
           if !addr.is_unspecified() {
             spec_dst = Some(IpAddr::V4(addr));
